Apply Numbrs changes to webrtc
1. Numbrs - Encapsulate camera1 and camera2 logic

Use RTCCameraInfo as the main interface to access any camera parameter
from camera1 and camera2.
Camera2 and 1 sessions can now use the same common model which is
RtcCameraInfo.

2. Numbrs - Add common interface to videoCapturer

Added needed methods to videoCapturer
and unified CameraSessionApi.

3. Numbrs - Remove unnecessary interfaces

4. Numbrs - Move CameraSessionApi to the java folder

Move CameraSessionApi to the java folder
so that it compiles.

5. Numbrs - Fix NullPointerException on camera1

6. Numbrs - Fix front facing for Camera1

7. Numbrs - Fix torch for Camera1

8. Numbrs - Replacing Template Preview with Record

Restores the original template used for video capturing.

9. Extend CameraCapturers to allow RtcCameraInfo

Also added a Makefile for automation.

10. Fix applyAutofocusFix

11. Apply image stretched flag

Apply image stretched flag to choose between TEMPLATE_RECORD (default)
and TEMPLATE_PREVIEW modes.

12. Refactor CameraCapturers

CameraCapturerCompact is the new class that handles both Camera1 and
Camera2 apis.

13. Skip black frames

Added an option to Camera2Session so that from the beginning of the
capturing black frames can be detected and skipped.

TicketId: PM-46454